requirements.txt
wheel
Pillow==8.1.2
pyperclip==1.8.2
PyQt5==5.15.4
PyQt5-sip==12.8.1
pytesseract==0.3.7
numpy==1.20.2
scipy
matplotlib
scikit-image==0.18.1
black
requests
beautifulsoup4
